requirements.txt
Django==2.2.6
Mastodon.py==1.4.6
